Change structure
The objects file is located in a folder now.
